feat: ViewModel 생성 및 책임 분리
